added whole_archive to source-libs
usually you want to let the linker decide which functions of a
library should be linked into the executable. this does not work
for some situations where a compile time plugin system registers
stuff with variables that are not used in the program. e.g. unit-tests.
the old cxxproject handled every library as whole-archive which led to
huge binaries ... now you can flag only the needed libraries as such.

Change-Id: I7a502668c4bc74fc6e5460f26494dfa48a1ba2fa